republican_calendar::date

Module consts

Constants for the Republican Calendar.

Constants

DAYS_IN_COMMON_YEAR

The number of days in a common year in the Republican Calendar.

DAYS_IN_LEAP_YEAR

The number of days in a leap year in the Republican Calendar.

DAYS_IN_MONTH

The number of days in a normal month in the Republican Calendar.

DAYS_IN_NORMAL_MONTHS

The number of days in the normal months of a common year in the Republican Calendar.

DAYS_IN_SANS_CULOTTIDES_COMMON_YEAR

The number of days in the Sans-culottides in a common year in the Republican Calendar.

DAYS_IN_SANS_CULOTTIDES_LEAP_YEAR

The number of days in the Sans-culottides in a leap year in the Republican Calendar.

FISRT_DAY

The first day of the Republican Calendar: 22 September 1792 in the Gregorian Calendar.

NORMAL_MONTHS_IN_YEAR

The number of months in a common year in the Republican Calendar (not counting Sans-culottides).

SANS_CULOTTIDES_MONTH_INDEX

The nominal index of the Sans-culottides month (13).