republican_calendar::date

Enum DayOfYear

pub enum DayOfYear {

    Raisin = 1,
    Safran = 2,
    Châtaigne = 3,
    Colchique = 4,
    Cheval = 5,
    Balsamine = 6,
    Carotte = 7,
    Amaranthe = 8,
    Panais = 9,
    Cuve = 10,
    PommeDeTerre = 11,
    Immortelle = 12,
    Potiron = 13,
    Réséda = 14,
    Âne = 15,
    BelleDeNuit = 16,
    Citrouille = 17,
    Sarrasin = 18,
    Tournesol = 19,
    Pressoir = 20,
    Chanvre = 21,
    Pêche = 22,
    Navet = 23,
    Amaryllis = 24,
    Bœuf = 25,
    Aubergine = 26,
    Piment = 27,
    Tomate = 28,
    Orge = 29,
    Tonneau = 30,
    Pomme = 31,
    Céleri = 32,
    Poire = 33,
    Betterave = 34,
    Oie = 35,
    Héliotrope = 36,
    Figue = 37,
    Scorsonère = 38,
    Alisier = 39,
    Charrue = 40,
    Salsifis = 41,
    Mâcre = 42,
    Topinambour = 43,
    Endive = 44,
    Dindon = 45,
    Chervis = 46,
    Cresson = 47,
    Dentelaire = 48,
    Grenade = 49,
    Herse = 50,
    Bacchante = 51,
    Azerole = 52,
    Garance = 53,
    Orange = 54,
    Faisan = 55,
    Pistache = 56,
    Macjonc = 57,
    Coing = 58,
    Cormier = 59,
    Rouleau = 60,
    Raiponce = 61,
    Turneps = 62,
    Chicorée = 63,
    Nèfle = 64,
    Cochon = 65,
    Mâche = 66,
    ChouFleur = 67,
    Miel = 68,
    Genièvre = 69,
    Pioche = 70,
    Cire = 71,
    Raifort = 72,
    Cèdre = 73,
    Sapin = 74,
    Chevreuil = 75,
    Ajonc = 76,
    Cyprès = 77,
    Lierre = 78,
    Sabine = 79,
    Hoyau = 80,
    ÉrableÀSucre = 81,
    Bruyère = 82,
    Roseau = 83,
    Oseille = 84,
    Grillon = 85,
    Pignon = 86,
    Liège = 87,
    Truffe = 88,
    Olive = 89,
    Pelle = 90,
    Tourbe = 91,
    Houille = 92,
    Bitume = 93,
    Soufre = 94,
    Chien = 95,
    Lave = 96,
    TerreVégétale = 97,
    Fumier = 98,
    Salpêtre = 99,
    Fléau = 100,
    Granit = 101,
    Argile = 102,
    Ardoise = 103,
    Grès = 104,
    Lapin = 105,
    Silex = 106,
    Marne = 107,
    PierreÀChaux = 108,
    Marbre = 109,
    Van = 110,
    PierreÀPlâtre = 111,
    Sel = 112,
    Fer = 113,
    Cuivre = 114,
    Chat = 115,
    Étain = 116,
    Plomb = 117,
    Zinc = 118,
    Mercure = 119,
    Crible = 120,
    Lauréole = 121,
    Mousse = 122,
    Fragon = 123,
    PerceNeige = 124,
    Taureau = 125,
    LaurierThym = 126,
    Amadouvier = 127,
    Mézéréon = 128,
    Peuplier = 129,
    Coignée = 130,
    Ellébore = 131,
    Brocoli = 132,
    Laurier = 133,
    Avelinier = 134,
    Vache = 135,
    Buis = 136,
    Lichen = 137,
    If = 138,
    Pulmonaire = 139,
    Serpette = 140,
    Thlaspi = 141,
    Thimelé = 142,
    Chiendent = 143,
    Trainasse = 144,
    Lièvre = 145,
    Guède = 146,
    Noisetier = 147,
    Cyclamen = 148,
    Chélidoine = 149,
    Traîneau = 150,
    Tussilage = 151,
    Cornouiller = 152,
    Violier = 153,
    Troène = 154,
    Bouc = 155,
    Asaret = 156,
    Alaterne = 157,
    Violette = 158,
    Marceau = 159,
    Bêche = 160,
    Narcisse = 161,
    Orme = 162,
    Fumeterre = 163,
    Vélar = 164,
    Chèvre = 165,
    Épinard = 166,
    Doronic = 167,
    Mouron = 168,
    Cerfeuil = 169,
    Cordeau = 170,
    Mandragore = 171,
    Persil = 172,
    Cochléaria = 173,
    Pâquerette = 174,
    Thon = 175,
    Pissenlit = 176,
    Sylvie = 177,
    Capillaire = 178,
    Frêne = 179,
    Plantoir = 180,
    Primevère = 181,
    Platane = 182,
    Asperge = 183,
    Tulipe = 184,
    Poule = 185,
    Bette = 186,
    Bouleau = 187,
    Jonquille = 188,
    Aulne = 189,
    Couvoir = 190,
    Pervenche = 191,
    Charme = 192,
    Morille = 193,
    Hêtre = 194,
    Abeille = 195,
    Laitue = 196,
    Mélèze = 197,
    Ciguë = 198,
    Radis = 199,
    Ruche = 200,
    Gainier = 201,
    Romaine = 202,
    Marronnier = 203,
    Roquette = 204,
    Pigeon = 205,
    Lilas = 206,
    Anémone = 207,
    Pensée = 208,
    Myrtille = 209,
    Greffoir = 210,
    Rose = 211,
    Chêne = 212,
    Fougère = 213,
    Aubépine = 214,
    Rossignol = 215,
    Ancolie = 216,
    Muguet = 217,
    Champignon = 218,
    Hyacinthe = 219,
    Râteau = 220,
    Rhubarbe = 221,
    Sainfoin = 222,
    BâtonDOr = 223,
    Chamerisier = 224,
    VerÀSoie = 225,
    Consoude = 226,
    Pimprenelle = 227,
    CorbeilleDOr = 228,
    Arroche = 229,
    Sarcloir = 230,
    Statice = 231,
    Fritillaire = 232,
    Bourrache = 233,
    Valériane = 234,
    Carpe = 235,
    Fusain = 236,
    Civette = 237,
    Buglosse = 238,
    Sénevé = 239,
    Houlette = 240,
    Luzerne = 241,
    Hémérocalle = 242,
    Trèfle = 243,
    Angélique = 244,
    Canard = 245,
    Mélisse = 246,
    Fromental = 247,
    Martagon = 248,
    Serpolet = 249,
    Faux = 250,
    Fraise = 251,
    Bétoine = 252,
    Pois = 253,
    Acacia = 254,
    Caille = 255,
    Œillet = 256,
    Sureau = 257,
    Pavot = 258,
    Tilleul = 259,
    Fourche = 260,
    Barbeau = 261,
    Camomille = 262,
    Chèvrefeuille = 263,
    CailleLait = 264,
    Tanche = 265,
    Jasmin = 266,
    Verveine = 267,
    Thym = 268,
    Pivoine = 269,
    Chariot = 270,
    Seigle = 271,
    Avoine = 272,
    Oignon = 273,
    Véronique = 274,
    Mulet = 275,
    Romarin = 276,
    Concombre = 277,
    Échalote = 278,
    Absinthe = 279,
    Faucille = 280,
    Coriandre = 281,
    Artichaut = 282,
    Girofle = 283,
    Lavande = 284,
    Chamois = 285,
    Tabac = 286,
    Groseille = 287,
    Gesse = 288,
    Cerise = 289,
    Parc = 290,
    Menthe = 291,
    Cumin = 292,
    Haricot = 293,
    Orcanète = 294,
    Pintade = 295,
    Sauge = 296,
    Ail = 297,
    Vesce = 298,
    Blé = 299,
    Chalémie = 300,
    Épeautre = 301,
    BouillonBlanc = 302,
    Melon = 303,
    Ivraie = 304,
    Bélier = 305,
    Prêle = 306,
    Armoise = 307,
    Carthame = 308,
    Mûre = 309,
    Arrosoir = 310,
    Panic = 311,
    Salicorne = 312,
    Abricot = 313,
    Basilic = 314,
    Brebis = 315,
    Guimauve = 316,
    Lin = 317,
    Amande = 318,
    Gentiane = 319,
    Écluse = 320,
    Carline = 321,
    Câprier = 322,
    Lentille = 323,
    Aunée = 324,
    Loutre = 325,
    Myrte = 326,
    Colza = 327,
    Lupin = 328,
    Coton = 329,
    Moulin = 330,
    Prune = 331,
    Millet = 332,
    Lycoperdon = 333,
    Escourgeon = 334,
    Saumon = 335,
    Tubéreuse = 336,
    Sucrion = 337,
    Apocyn = 338,
    Réglisse = 339,
    Échelle = 340,
    Pastèque = 341,
    Fenouil = 342,
    ÉpineVinette = 343,
    Noix = 344,
    Truite = 345,
    Citron = 346,
    Cardère = 347,
    Nerprun = 348,
    Tagette = 349,
    Hotte = 350,
    Églantier = 351,
    Noisette = 352,
    Houblon = 353,
    Sorgho = 354,
    Écrevisse = 355,
    Bigarade = 356,
    VergeDOr = 357,
    Maïs = 358,
    Marron = 359,
    Panier = 360,
    LaFêteDeLaVertu = 361,
    LaFêteDuGénie = 362,
    LaFêteDuTravail = 363,
    LaFêteDeLOpinion = 364,
    LaFêteDesRécompenses = 365,
    LaFêteDeLaRévolution = 366,

}

Days in a year of the Republican Calendar

Variants

Raisin = 1

Raisin, day 1 of Vendémiaire, day 1 of Autumn, day 1 of the year.

Safran = 2

Safran, day 2 of Vendémiaire, day 2 of Autumn, day 2 of the year.

Châtaigne = 3

Châtaigne, day 3 of Vendémiaire, day 3 of Autumn, day 3 of the year.

Colchique = 4

Colchique, day 4 of Vendémiaire, day 4 of Autumn, day 4 of the year.

Cheval = 5

Cheval, day 5 of Vendémiaire, day 5 of Autumn, day 5 of the year.

Balsamine = 6

Balsamine, day 6 of Vendémiaire, day 6 of Autumn, day 6 of the year.

Carotte = 7

Carotte, day 7 of Vendémiaire, day 7 of Autumn, day 7 of the year.

Amaranthe = 8

Amaranthe, day 8 of Vendémiaire, day 8 of Autumn, day 8 of the year.

Panais = 9

Panais, day 9 of Vendémiaire, day 9 of Autumn, day 9 of the year.

Cuve = 10

Cuve, day 10 of Vendémiaire, day 10 of Autumn, day 10 of the year.

PommeDeTerre = 11

Pomme de terre, day 11 of Vendémiaire, day 11 of Autumn, day 11 of the year.

Immortelle = 12

Immortelle, day 12 of Vendémiaire, day 12 of Autumn, day 12 of the year.

Potiron = 13

Potiron, day 13 of Vendémiaire, day 13 of Autumn, day 13 of the year.

Réséda = 14

Réséda, day 14 of Vendémiaire, day 14 of Autumn, day 14 of the year.

Âne = 15

Âne, day 15 of Vendémiaire, day 15 of Autumn, day 15 of the year.

BelleDeNuit = 16

Belle de nuit, day 16 of Vendémiaire, day 16 of Autumn, day 16 of the year.

Citrouille = 17

Citrouille, day 17 of Vendémiaire, day 17 of Autumn, day 17 of the year.

Sarrasin = 18

Sarrasin, day 18 of Vendémiaire, day 18 of Autumn, day 18 of the year.

Tournesol = 19

Tournesol, day 19 of Vendémiaire, day 19 of Autumn, day 19 of the year.

Pressoir = 20

Pressoir, day 20 of Vendémiaire, day 20 of Autumn, day 20 of the year.

Chanvre = 21

Chanvre, day 21 of Vendémiaire, day 21 of Autumn, day 21 of the year.

Pêche = 22

Pêche, day 22 of Vendémiaire, day 22 of Autumn, day 22 of the year.

Navet = 23

Navet, day 23 of Vendémiaire, day 23 of Autumn, day 23 of the year.

Amaryllis = 24

Amaryllis, day 24 of Vendémiaire, day 24 of Autumn, day 24 of the year.

Bœuf = 25

Bœuf, day 25 of Vendémiaire, day 25 of Autumn, day 25 of the year.

Aubergine = 26

Aubergine, day 26 of Vendémiaire, day 26 of Autumn, day 26 of the year.

Piment = 27

Piment, day 27 of Vendémiaire, day 27 of Autumn, day 27 of the year.

Tomate = 28

Tomate, day 28 of Vendémiaire, day 28 of Autumn, day 28 of the year.

Orge = 29

Orge, day 29 of Vendémiaire, day 29 of Autumn, day 29 of the year.

Tonneau = 30

Tonneau, day 30 of Vendémiaire, day 30 of Autumn, day 30 of the year.

Pomme = 31

Pomme, day 1 of Brumaire, day 31 of Autumn, day 31 of the year.

Céleri = 32

Céleri, day 2 of Brumaire, day 32 of Autumn, day 32 of the year.

Poire = 33

Poire, day 3 of Brumaire, day 33 of Autumn, day 33 of the year.

Betterave = 34

Betterave, day 4 of Brumaire, day 34 of Autumn, day 34 of the year.

Oie = 35

Oie, day 5 of Brumaire, day 35 of Autumn, day 35 of the year.

Héliotrope = 36

Héliotrope, day 6 of Brumaire, day 36 of Autumn, day 36 of the year.

Figue = 37

Figue, day 7 of Brumaire, day 37 of Autumn, day 37 of the year.

Scorsonère = 38

Scorsonère, day 8 of Brumaire, day 38 of Autumn, day 38 of the year.

Alisier = 39

Alisier, day 9 of Brumaire, day 39 of Autumn, day 39 of the year.

Charrue = 40

Charrue, day 10 of Brumaire, day 40 of Autumn, day 40 of the year.

Salsifis = 41

Salsifis, day 11 of Brumaire, day 41 of Autumn, day 41 of the year.

Mâcre = 42

Mâcre, day 12 of Brumaire, day 42 of Autumn, day 42 of the year.

Topinambour = 43

Topinambour, day 13 of Brumaire, day 43 of Autumn, day 43 of the year.

Endive = 44

Endive, day 14 of Brumaire, day 44 of Autumn, day 44 of the year.

Dindon = 45

Dindon, day 15 of Brumaire, day 45 of Autumn, day 45 of the year.

Chervis = 46

Chervis, day 16 of Brumaire, day 46 of Autumn, day 46 of the year.

Cresson = 47

Cresson, day 17 of Brumaire, day 47 of Autumn, day 47 of the year.

Dentelaire = 48

Dentelaire, day 18 of Brumaire, day 48 of Autumn, day 48 of the year.

Grenade = 49

Grenade, day 19 of Brumaire, day 49 of Autumn, day 49 of the year.

Herse = 50

Herse, day 20 of Brumaire, day 50 of Autumn, day 50 of the year.

Bacchante = 51

Bacchante, day 21 of Brumaire, day 51 of Autumn, day 51 of the year.

Azerole = 52

Azerole, day 22 of Brumaire, day 52 of Autumn, day 52 of the year.

Garance = 53

Garance, day 23 of Brumaire, day 53 of Autumn, day 53 of the year.

Orange = 54

Orange, day 24 of Brumaire, day 54 of Autumn, day 54 of the year.

Faisan = 55

Faisan, day 25 of Brumaire, day 55 of Autumn, day 55 of the year.

Pistache = 56

Pistache, day 26 of Brumaire, day 56 of Autumn, day 56 of the year.

Macjonc = 57

Macjonc, day 27 of Brumaire, day 57 of Autumn, day 57 of the year.

Coing = 58

Coing, day 28 of Brumaire, day 58 of Autumn, day 58 of the year.

Cormier = 59

Cormier, day 29 of Brumaire, day 59 of Autumn, day 59 of the year.

Rouleau = 60

Rouleau, day 30 of Brumaire, day 60 of Autumn, day 60 of the year.

Raiponce = 61

Raiponce, day 1 of Frimaire, day 61 of Autumn, day 61 of the year.

Turneps = 62

Turneps, day 2 of Frimaire, day 62 of Autumn, day 62 of the year.

Chicorée = 63

Chicorée, day 3 of Frimaire, day 63 of Autumn, day 63 of the year.

Nèfle = 64

Nèfle, day 4 of Frimaire, day 64 of Autumn, day 64 of the year.

Cochon = 65

Cochon, day 5 of Frimaire, day 65 of Autumn, day 65 of the year.

Mâche = 66

Mâche, day 6 of Frimaire, day 66 of Autumn, day 66 of the year.

ChouFleur = 67

Chou-fleur, day 7 of Frimaire, day 67 of Autumn, day 67 of the year.

Miel = 68

Miel, day 8 of Frimaire, day 68 of Autumn, day 68 of the year.

Genièvre = 69

Genièvre, day 9 of Frimaire, day 69 of Autumn, day 69 of the year.

Pioche = 70

Pioche, day 10 of Frimaire, day 70 of Autumn, day 70 of the year.

Cire = 71

Cire, day 11 of Frimaire, day 71 of Autumn, day 71 of the year.

Raifort = 72

Raifort, day 12 of Frimaire, day 72 of Autumn, day 72 of the year.

Cèdre = 73

Cèdre, day 13 of Frimaire, day 73 of Autumn, day 73 of the year.

Sapin = 74

Sapin, day 14 of Frimaire, day 74 of Autumn, day 74 of the year.

Chevreuil = 75

Chevreuil, day 15 of Frimaire, day 75 of Autumn, day 75 of the year.

Ajonc = 76

Ajonc, day 16 of Frimaire, day 76 of Autumn, day 76 of the year.

Cyprès = 77

Cyprès, day 17 of Frimaire, day 77 of Autumn, day 77 of the year.

Lierre = 78

Lierre, day 18 of Frimaire, day 78 of Autumn, day 78 of the year.

Sabine = 79

Sabine, day 19 of Frimaire, day 79 of Autumn, day 79 of the year.

Hoyau = 80

Hoyau, day 20 of Frimaire, day 80 of Autumn, day 80 of the year.

ÉrableÀSucre = 81

Érable à sucre, day 21 of Frimaire, day 81 of Autumn, day 81 of the year.

Bruyère = 82

Bruyère, day 22 of Frimaire, day 82 of Autumn, day 82 of the year.

Roseau = 83

Roseau, day 23 of Frimaire, day 83 of Autumn, day 83 of the year.

Oseille = 84

Oseille, day 24 of Frimaire, day 84 of Autumn, day 84 of the year.

Grillon = 85

Grillon, day 25 of Frimaire, day 85 of Autumn, day 85 of the year.

Pignon = 86

Pignon, day 26 of Frimaire, day 86 of Autumn, day 86 of the year.

Liège = 87

Liège, day 27 of Frimaire, day 87 of Autumn, day 87 of the year.

Truffe = 88

Truffe, day 28 of Frimaire, day 88 of Autumn, day 88 of the year.

Olive = 89

Olive, day 29 of Frimaire, day 89 of Autumn, day 89 of the year.

Pelle = 90

Pelle, day 30 of Frimaire, day 90 of Autumn, day 90 of the year.

Tourbe = 91

Tourbe, day 1 of Nivôse, day 1 of Winter, day 91 of the year.

Houille = 92

Houille, day 2 of Nivôse, day 2 of Winter, day 92 of the year.

Bitume = 93

Bitume, day 3 of Nivôse, day 3 of Winter, day 93 of the year.

Soufre = 94

Soufre, day 4 of Nivôse, day 4 of Winter, day 94 of the year.

Chien = 95

Chien, day 5 of Nivôse, day 5 of Winter, day 95 of the year.

Lave = 96

Lave, day 6 of Nivôse, day 6 of Winter, day 96 of the year.

TerreVégétale = 97

Terre végétale, day 7 of Nivôse, day 7 of Winter, day 97 of the year.

Fumier = 98

Fumier, day 8 of Nivôse, day 8 of Winter, day 98 of the year.

Salpêtre = 99

Salpêtre, day 9 of Nivôse, day 9 of Winter, day 99 of the year.

Fléau = 100

Fléau, day 10 of Nivôse, day 10 of Winter, day 100 of the year.

Granit = 101

Granit, day 11 of Nivôse, day 11 of Winter, day 101 of the year.

Argile = 102

Argile, day 12 of Nivôse, day 12 of Winter, day 102 of the year.

Ardoise = 103

Ardoise, day 13 of Nivôse, day 13 of Winter, day 103 of the year.

Grès = 104

Grès, day 14 of Nivôse, day 14 of Winter, day 104 of the year.

Lapin = 105

Lapin, day 15 of Nivôse, day 15 of Winter, day 105 of the year.

Silex = 106

Silex, day 16 of Nivôse, day 16 of Winter, day 106 of the year.

Marne = 107

Marne, day 17 of Nivôse, day 17 of Winter, day 107 of the year.

PierreÀChaux = 108

Pierre à chaux, day 18 of Nivôse, day 18 of Winter, day 108 of the year.

Marbre = 109

Marbre, day 19 of Nivôse, day 19 of Winter, day 109 of the year.

Van = 110

Van, day 20 of Nivôse, day 20 of Winter, day 110 of the year.

PierreÀPlâtre = 111

Pierre à plâtre, day 21 of Nivôse, day 21 of Winter, day 111 of the year.

Sel = 112

Sel, day 22 of Nivôse, day 22 of Winter, day 112 of the year.

Fer = 113

Fer, day 23 of Nivôse, day 23 of Winter, day 113 of the year.

Cuivre = 114

Cuivre, day 24 of Nivôse, day 24 of Winter, day 114 of the year.

Chat = 115

Chat, day 25 of Nivôse, day 25 of Winter, day 115 of the year.

Étain = 116

Étain, day 26 of Nivôse, day 26 of Winter, day 116 of the year.

Plomb = 117

Plomb, day 27 of Nivôse, day 27 of Winter, day 117 of the year.

Zinc = 118

Zinc, day 28 of Nivôse, day 28 of Winter, day 118 of the year.

Mercure = 119

Mercure, day 29 of Nivôse, day 29 of Winter, day 119 of the year.

Crible = 120

Crible, day 30 of Nivôse, day 30 of Winter, day 120 of the year.

Lauréole = 121

Lauréole, day 1 of Pluviôse, day 31 of Winter, day 121 of the year.

Mousse = 122

Mousse, day 2 of Pluviôse, day 32 of Winter, day 122 of the year.

Fragon = 123

Fragon, day 3 of Pluviôse, day 33 of Winter, day 123 of the year.

PerceNeige = 124

Perce-neige, day 4 of Pluviôse, day 34 of Winter, day 124 of the year.

Taureau = 125

Taureau, day 5 of Pluviôse, day 35 of Winter, day 125 of the year.

LaurierThym = 126

Laurier-thym, day 6 of Pluviôse, day 36 of Winter, day 126 of the year.

Amadouvier = 127

Amadouvier, day 7 of Pluviôse, day 37 of Winter, day 127 of the year.

Mézéréon = 128

Mézéréon, day 8 of Pluviôse, day 38 of Winter, day 128 of the year.

Peuplier = 129

Peuplier, day 9 of Pluviôse, day 39 of Winter, day 129 of the year.

Coignée = 130

Coignée, day 10 of Pluviôse, day 40 of Winter, day 130 of the year.

Ellébore = 131

Ellébore, day 11 of Pluviôse, day 41 of Winter, day 131 of the year.

Brocoli = 132

Brocoli, day 12 of Pluviôse, day 42 of Winter, day 132 of the year.

Laurier = 133

Laurier, day 13 of Pluviôse, day 43 of Winter, day 133 of the year.

Avelinier = 134

Avelinier, day 14 of Pluviôse, day 44 of Winter, day 134 of the year.

Vache = 135

Vache, day 15 of Pluviôse, day 45 of Winter, day 135 of the year.

Buis = 136

Buis, day 16 of Pluviôse, day 46 of Winter, day 136 of the year.

Lichen = 137

Lichen, day 17 of Pluviôse, day 47 of Winter, day 137 of the year.

If = 138

If, day 18 of Pluviôse, day 48 of Winter, day 138 of the year.

Pulmonaire = 139

Pulmonaire, day 19 of Pluviôse, day 49 of Winter, day 139 of the year.

Serpette = 140

Serpette, day 20 of Pluviôse, day 50 of Winter, day 140 of the year.

Thlaspi = 141

Thlaspi, day 21 of Pluviôse, day 51 of Winter, day 141 of the year.

Thimelé = 142

Thimelé, day 22 of Pluviôse, day 52 of Winter, day 142 of the year.

Chiendent = 143

Chiendent, day 23 of Pluviôse, day 53 of Winter, day 143 of the year.

Trainasse = 144

Trainasse, day 24 of Pluviôse, day 54 of Winter, day 144 of the year.

Lièvre = 145

Lièvre, day 25 of Pluviôse, day 55 of Winter, day 145 of the year.

Guède = 146

Guède, day 26 of Pluviôse, day 56 of Winter, day 146 of the year.

Noisetier = 147

Noisetier, day 27 of Pluviôse, day 57 of Winter, day 147 of the year.

Cyclamen = 148

Cyclamen, day 28 of Pluviôse, day 58 of Winter, day 148 of the year.

Chélidoine = 149

Chélidoine, day 29 of Pluviôse, day 59 of Winter, day 149 of the year.

Traîneau = 150

Traîneau, day 30 of Pluviôse, day 60 of Winter, day 150 of the year.

Tussilage = 151

Tussilage, day 1 of Ventôse, day 61 of Winter, day 151 of the year.

Cornouiller = 152

Cornouiller, day 2 of Ventôse, day 62 of Winter, day 152 of the year.

Violier = 153

Violier, day 3 of Ventôse, day 63 of Winter, day 153 of the year.

Troène = 154

Troène, day 4 of Ventôse, day 64 of Winter, day 154 of the year.

Bouc = 155

Bouc, day 5 of Ventôse, day 65 of Winter, day 155 of the year.

Asaret = 156

Asaret, day 6 of Ventôse, day 66 of Winter, day 156 of the year.

Alaterne = 157

Alaterne, day 7 of Ventôse, day 67 of Winter, day 157 of the year.

Violette = 158

Violette, day 8 of Ventôse, day 68 of Winter, day 158 of the year.

Marceau = 159

Marceau, day 9 of Ventôse, day 69 of Winter, day 159 of the year.

Bêche = 160

Bêche, day 10 of Ventôse, day 70 of Winter, day 160 of the year.

Narcisse = 161

Narcisse, day 11 of Ventôse, day 71 of Winter, day 161 of the year.

Orme = 162

Orme, day 12 of Ventôse, day 72 of Winter, day 162 of the year.

Fumeterre = 163

Fumeterre, day 13 of Ventôse, day 73 of Winter, day 163 of the year.

Vélar = 164

Vélar, day 14 of Ventôse, day 74 of Winter, day 164 of the year.

Chèvre = 165

Chèvre, day 15 of Ventôse, day 75 of Winter, day 165 of the year.

Épinard = 166

Épinard, day 16 of Ventôse, day 76 of Winter, day 166 of the year.

Doronic = 167

Doronic, day 17 of Ventôse, day 77 of Winter, day 167 of the year.

Mouron = 168

Mouron, day 18 of Ventôse, day 78 of Winter, day 168 of the year.

Cerfeuil = 169

Cerfeuil, day 19 of Ventôse, day 79 of Winter, day 169 of the year.

Cordeau = 170

Cordeau, day 20 of Ventôse, day 80 of Winter, day 170 of the year.

Mandragore = 171

Mandragore, day 21 of Ventôse, day 81 of Winter, day 171 of the year.

Persil = 172

Persil, day 22 of Ventôse, day 82 of Winter, day 172 of the year.

Cochléaria = 173

Cochléaria, day 23 of Ventôse, day 83 of Winter, day 173 of the year.

Pâquerette = 174

Pâquerette, day 24 of Ventôse, day 84 of Winter, day 174 of the year.

Thon = 175

Thon, day 25 of Ventôse, day 85 of Winter, day 175 of the year.

Pissenlit = 176

Pissenlit, day 26 of Ventôse, day 86 of Winter, day 176 of the year.

Sylvie = 177

Sylvie, day 27 of Ventôse, day 87 of Winter, day 177 of the year.

Capillaire = 178

Capillaire, day 28 of Ventôse, day 88 of Winter, day 178 of the year.

Frêne = 179

Frêne, day 29 of Ventôse, day 89 of Winter, day 179 of the year.

Plantoir = 180

Plantoir, day 30 of Ventôse, day 90 of Winter, day 180 of the year.

Primevère = 181

Primevère, day 1 of Germinal, day 1 of Spring, day 181 of the year.

Platane = 182

Platane, day 2 of Germinal, day 2 of Spring, day 182 of the year.

Asperge = 183

Asperge, day 3 of Germinal, day 3 of Spring, day 183 of the year.

Tulipe = 184

Tulipe, day 4 of Germinal, day 4 of Spring, day 184 of the year.

Poule = 185

Poule, day 5 of Germinal, day 5 of Spring, day 185 of the year.

Bette = 186

Bette, day 6 of Germinal, day 6 of Spring, day 186 of the year.

Bouleau = 187

Bouleau, day 7 of Germinal, day 7 of Spring, day 187 of the year.

Jonquille = 188

Jonquille, day 8 of Germinal, day 8 of Spring, day 188 of the year.

Aulne = 189

Aulne, day 9 of Germinal, day 9 of Spring, day 189 of the year.

Couvoir = 190

Couvoir, day 10 of Germinal, day 10 of Spring, day 190 of the year.

Pervenche = 191

Pervenche, day 11 of Germinal, day 11 of Spring, day 191 of the year.

Charme = 192

Charme, day 12 of Germinal, day 12 of Spring, day 192 of the year.

Morille = 193

Morille, day 13 of Germinal, day 13 of Spring, day 193 of the year.

Hêtre = 194

Hêtre, day 14 of Germinal, day 14 of Spring, day 194 of the year.

Abeille = 195

Abeille, day 15 of Germinal, day 15 of Spring, day 195 of the year.

Laitue = 196

Laitue, day 16 of Germinal, day 16 of Spring, day 196 of the year.

Mélèze = 197

Mélèze, day 17 of Germinal, day 17 of Spring, day 197 of the year.

Ciguë = 198

Ciguë, day 18 of Germinal, day 18 of Spring, day 198 of the year.

Radis = 199

Radis, day 19 of Germinal, day 19 of Spring, day 199 of the year.

Ruche = 200

Ruche, day 20 of Germinal, day 20 of Spring, day 200 of the year.

Gainier = 201

Gainier, day 21 of Germinal, day 21 of Spring, day 201 of the year.

Romaine = 202

Romaine, day 22 of Germinal, day 22 of Spring, day 202 of the year.

Marronnier = 203

Marronnier, day 23 of Germinal, day 23 of Spring, day 203 of the year.

Roquette = 204

Roquette, day 24 of Germinal, day 24 of Spring, day 204 of the year.

Pigeon = 205

Pigeon, day 25 of Germinal, day 25 of Spring, day 205 of the year.

Lilas = 206

Lilas, day 26 of Germinal, day 26 of Spring, day 206 of the year.

Anémone = 207

Anémone, day 27 of Germinal, day 27 of Spring, day 207 of the year.

Pensée = 208

Pensée, day 28 of Germinal, day 28 of Spring, day 208 of the year.

Myrtille = 209

Myrtille, day 29 of Germinal, day 29 of Spring, day 209 of the year.

Greffoir = 210

Greffoir, day 30 of Germinal, day 30 of Spring, day 210 of the year.

Rose = 211

Rose, day 1 of Floréal, day 31 of Spring, day 211 of the year.

Chêne = 212

Chêne, day 2 of Floréal, day 32 of Spring, day 212 of the year.

Fougère = 213

Fougère, day 3 of Floréal, day 33 of Spring, day 213 of the year.

Aubépine = 214

Aubépine, day 4 of Floréal, day 34 of Spring, day 214 of the year.

Rossignol = 215

Rossignol, day 5 of Floréal, day 35 of Spring, day 215 of the year.

Ancolie = 216

Ancolie, day 6 of Floréal, day 36 of Spring, day 216 of the year.

Muguet = 217

Muguet, day 7 of Floréal, day 37 of Spring, day 217 of the year.

Champignon = 218

Champignon, day 8 of Floréal, day 38 of Spring, day 218 of the year.

Hyacinthe = 219

Hyacinthe, day 9 of Floréal, day 39 of Spring, day 219 of the year.

Râteau = 220

Râteau, day 10 of Floréal, day 40 of Spring, day 220 of the year.

Rhubarbe = 221

Rhubarbe, day 11 of Floréal, day 41 of Spring, day 221 of the year.

Sainfoin = 222

Sainfoin, day 12 of Floréal, day 42 of Spring, day 222 of the year.

BâtonDOr = 223

Bâton d’or, day 13 of Floréal, day 43 of Spring, day 223 of the year.

Chamerisier = 224

Chamerisier, day 14 of Floréal, day 44 of Spring, day 224 of the year.

VerÀSoie = 225

Ver à soie, day 15 of Floréal, day 45 of Spring, day 225 of the year.

Consoude = 226

Consoude, day 16 of Floréal, day 46 of Spring, day 226 of the year.

Pimprenelle = 227

Pimprenelle, day 17 of Floréal, day 47 of Spring, day 227 of the year.

CorbeilleDOr = 228

Corbeille d’or, day 18 of Floréal, day 48 of Spring, day 228 of the year.

Arroche = 229

Arroche, day 19 of Floréal, day 49 of Spring, day 229 of the year.

Sarcloir = 230

Sarcloir, day 20 of Floréal, day 50 of Spring, day 230 of the year.

Statice = 231

Statice, day 21 of Floréal, day 51 of Spring, day 231 of the year.

Fritillaire = 232

Fritillaire, day 22 of Floréal, day 52 of Spring, day 232 of the year.

Bourrache = 233

Bourrache, day 23 of Floréal, day 53 of Spring, day 233 of the year.

Valériane = 234

Valériane, day 24 of Floréal, day 54 of Spring, day 234 of the year.

Carpe = 235

Carpe, day 25 of Floréal, day 55 of Spring, day 235 of the year.

Fusain = 236

Fusain, day 26 of Floréal, day 56 of Spring, day 236 of the year.

Civette = 237

Civette, day 27 of Floréal, day 57 of Spring, day 237 of the year.

Buglosse = 238

Buglosse, day 28 of Floréal, day 58 of Spring, day 238 of the year.

Sénevé = 239

Sénevé, day 29 of Floréal, day 59 of Spring, day 239 of the year.

Houlette = 240

Houlette, day 30 of Floréal, day 60 of Spring, day 240 of the year.

Luzerne = 241

Luzerne, day 1 of Prairial, day 61 of Spring, day 241 of the year.

Hémérocalle = 242

Hémérocalle, day 2 of Prairial, day 62 of Spring, day 242 of the year.

Trèfle = 243

Trèfle, day 3 of Prairial, day 63 of Spring, day 243 of the year.

Angélique = 244

Angélique, day 4 of Prairial, day 64 of Spring, day 244 of the year.

Canard = 245

Canard, day 5 of Prairial, day 65 of Spring, day 245 of the year.

Mélisse = 246

Mélisse, day 6 of Prairial, day 66 of Spring, day 246 of the year.

Fromental = 247

Fromental, day 7 of Prairial, day 67 of Spring, day 247 of the year.

Martagon = 248

Martagon, day 8 of Prairial, day 68 of Spring, day 248 of the year.

Serpolet = 249

Serpolet, day 9 of Prairial, day 69 of Spring, day 249 of the year.

Faux = 250

Faux, day 10 of Prairial, day 70 of Spring, day 250 of the year.

Fraise = 251

Fraise, day 11 of Prairial, day 71 of Spring, day 251 of the year.

Bétoine = 252

Bétoine, day 12 of Prairial, day 72 of Spring, day 252 of the year.

Pois = 253

Pois, day 13 of Prairial, day 73 of Spring, day 253 of the year.

Acacia = 254

Acacia, day 14 of Prairial, day 74 of Spring, day 254 of the year.

Caille = 255

Caille, day 15 of Prairial, day 75 of Spring, day 255 of the year.

Œillet = 256

Œillet, day 16 of Prairial, day 76 of Spring, day 256 of the year.

Sureau = 257

Sureau, day 17 of Prairial, day 77 of Spring, day 257 of the year.

Pavot = 258

Pavot, day 18 of Prairial, day 78 of Spring, day 258 of the year.

Tilleul = 259

Tilleul, day 19 of Prairial, day 79 of Spring, day 259 of the year.

Fourche = 260

Fourche, day 20 of Prairial, day 80 of Spring, day 260 of the year.

Barbeau = 261

Barbeau, day 21 of Prairial, day 81 of Spring, day 261 of the year.

Camomille = 262

Camomille, day 22 of Prairial, day 82 of Spring, day 262 of the year.

Chèvrefeuille = 263

Chèvrefeuille, day 23 of Prairial, day 83 of Spring, day 263 of the year.

CailleLait = 264

Caille-lait, day 24 of Prairial, day 84 of Spring, day 264 of the year.

Tanche = 265

Tanche, day 25 of Prairial, day 85 of Spring, day 265 of the year.

Jasmin = 266

Jasmin, day 26 of Prairial, day 86 of Spring, day 266 of the year.

Verveine = 267

Verveine, day 27 of Prairial, day 87 of Spring, day 267 of the year.

Thym = 268

Thym, day 28 of Prairial, day 88 of Spring, day 268 of the year.

Pivoine = 269

Pivoine, day 29 of Prairial, day 89 of Spring, day 269 of the year.

Chariot = 270

Chariot, day 30 of Prairial, day 90 of Spring, day 270 of the year.

Seigle = 271

Seigle, day 1 of Messidor, day 1 of Summer, day 271 of the year.

Avoine = 272

Avoine, day 2 of Messidor, day 2 of Summer, day 272 of the year.

Oignon = 273

Oignon, day 3 of Messidor, day 3 of Summer, day 273 of the year.

Véronique = 274

Véronique, day 4 of Messidor, day 4 of Summer, day 274 of the year.

Mulet = 275

Mulet, day 5 of Messidor, day 5 of Summer, day 275 of the year.

Romarin = 276

Romarin, day 6 of Messidor, day 6 of Summer, day 276 of the year.

Concombre = 277

Concombre, day 7 of Messidor, day 7 of Summer, day 277 of the year.

Échalote = 278

Échalote, day 8 of Messidor, day 8 of Summer, day 278 of the year.

Absinthe = 279

Absinthe, day 9 of Messidor, day 9 of Summer, day 279 of the year.

Faucille = 280

Faucille, day 10 of Messidor, day 10 of Summer, day 280 of the year.

Coriandre = 281

Coriandre, day 11 of Messidor, day 11 of Summer, day 281 of the year.

Artichaut = 282

Artichaut, day 12 of Messidor, day 12 of Summer, day 282 of the year.

Girofle = 283

Girofle, day 13 of Messidor, day 13 of Summer, day 283 of the year.

Lavande = 284

Lavande, day 14 of Messidor, day 14 of Summer, day 284 of the year.

Chamois = 285

Chamois, day 15 of Messidor, day 15 of Summer, day 285 of the year.

Tabac = 286

Tabac, day 16 of Messidor, day 16 of Summer, day 286 of the year.

Groseille = 287

Groseille, day 17 of Messidor, day 17 of Summer, day 287 of the year.

Gesse = 288

Gesse, day 18 of Messidor, day 18 of Summer, day 288 of the year.

Cerise = 289

Cerise, day 19 of Messidor, day 19 of Summer, day 289 of the year.

Parc = 290

Parc, day 20 of Messidor, day 20 of Summer, day 290 of the year.

Menthe = 291

Menthe, day 21 of Messidor, day 21 of Summer, day 291 of the year.

Cumin = 292

Cumin, day 22 of Messidor, day 22 of Summer, day 292 of the year.

Haricot = 293

Haricot, day 23 of Messidor, day 23 of Summer, day 293 of the year.

Orcanète = 294

Orcanète, day 24 of Messidor, day 24 of Summer, day 294 of the year.

Pintade = 295

Pintade, day 25 of Messidor, day 25 of Summer, day 295 of the year.

Sauge = 296

Sauge, day 26 of Messidor, day 26 of Summer, day 296 of the year.

Ail = 297

Ail, day 27 of Messidor, day 27 of Summer, day 297 of the year.

Vesce = 298

Vesce, day 28 of Messidor, day 28 of Summer, day 298 of the year.

Blé = 299

Blé, day 29 of Messidor, day 29 of Summer, day 299 of the year.

Chalémie = 300

Chalémie, day 30 of Messidor, day 30 of Summer, day 300 of the year.

Épeautre = 301

Épeautre, day 1 of Thermidor, day 31 of Summer, day 301 of the year.

BouillonBlanc = 302

Bouillon blanc, day 2 of Thermidor, day 32 of Summer, day 302 of the year.

Melon = 303

Melon, day 3 of Thermidor, day 33 of Summer, day 303 of the year.

Ivraie = 304

Ivraie, day 4 of Thermidor, day 34 of Summer, day 304 of the year.

Bélier = 305

Bélier, day 5 of Thermidor, day 35 of Summer, day 305 of the year.

Prêle = 306

Prêle, day 6 of Thermidor, day 36 of Summer, day 306 of the year.

Armoise = 307

Armoise, day 7 of Thermidor, day 37 of Summer, day 307 of the year.

Carthame = 308

Carthame, day 8 of Thermidor, day 38 of Summer, day 308 of the year.

Mûre = 309

Mûre, day 9 of Thermidor, day 39 of Summer, day 309 of the year.

Arrosoir = 310

Arrosoir, day 10 of Thermidor, day 40 of Summer, day 310 of the year.

Panic = 311

Panic, day 11 of Thermidor, day 41 of Summer, day 311 of the year.

Salicorne = 312

Salicorne, day 12 of Thermidor, day 42 of Summer, day 312 of the year.

Abricot = 313

Abricot, day 13 of Thermidor, day 43 of Summer, day 313 of the year.

Basilic = 314

Basilic, day 14 of Thermidor, day 44 of Summer, day 314 of the year.

Brebis = 315

Brebis, day 15 of Thermidor, day 45 of Summer, day 315 of the year.

Guimauve = 316

Guimauve, day 16 of Thermidor, day 46 of Summer, day 316 of the year.

Lin = 317

Lin, day 17 of Thermidor, day 47 of Summer, day 317 of the year.

Amande = 318

Amande, day 18 of Thermidor, day 48 of Summer, day 318 of the year.

Gentiane = 319

Gentiane, day 19 of Thermidor, day 49 of Summer, day 319 of the year.

Écluse = 320

Écluse, day 20 of Thermidor, day 50 of Summer, day 320 of the year.

Carline = 321

Carline, day 21 of Thermidor, day 51 of Summer, day 321 of the year.

Câprier = 322

Câprier, day 22 of Thermidor, day 52 of Summer, day 322 of the year.

Lentille = 323

Lentille, day 23 of Thermidor, day 53 of Summer, day 323 of the year.

Aunée = 324

Aunée, day 24 of Thermidor, day 54 of Summer, day 324 of the year.

Loutre = 325

Loutre, day 25 of Thermidor, day 55 of Summer, day 325 of the year.

Myrte = 326

Myrte, day 26 of Thermidor, day 56 of Summer, day 326 of the year.

Colza = 327

Colza, day 27 of Thermidor, day 57 of Summer, day 327 of the year.

Lupin = 328

Lupin, day 28 of Thermidor, day 58 of Summer, day 328 of the year.

Coton = 329

Coton, day 29 of Thermidor, day 59 of Summer, day 329 of the year.

Moulin = 330

Moulin, day 30 of Thermidor, day 60 of Summer, day 330 of the year.

Prune = 331

Prune, day 1 of Fructidor, day 61 of Summer, day 331 of the year.

Millet = 332

Millet, day 2 of Fructidor, day 62 of Summer, day 332 of the year.

Lycoperdon = 333

Lycoperdon, day 3 of Fructidor, day 63 of Summer, day 333 of the year.

Escourgeon = 334

Escourgeon, day 4 of Fructidor, day 64 of Summer, day 334 of the year.

Saumon = 335

Saumon, day 5 of Fructidor, day 65 of Summer, day 335 of the year.

Tubéreuse = 336

Tubéreuse, day 6 of Fructidor, day 66 of Summer, day 336 of the year.

Sucrion = 337

Sucrion, day 7 of Fructidor, day 67 of Summer, day 337 of the year.

Apocyn = 338

Apocyn, day 8 of Fructidor, day 68 of Summer, day 338 of the year.

Réglisse = 339

Réglisse, day 9 of Fructidor, day 69 of Summer, day 339 of the year.

Échelle = 340

Échelle, day 10 of Fructidor, day 70 of Summer, day 340 of the year.

Pastèque = 341

Pastèque, day 11 of Fructidor, day 71 of Summer, day 341 of the year.

Fenouil = 342

Fenouil, day 12 of Fructidor, day 72 of Summer, day 342 of the year.

ÉpineVinette = 343

Épine vinette, day 13 of Fructidor, day 73 of Summer, day 343 of the year.

Noix = 344

Noix, day 14 of Fructidor, day 74 of Summer, day 344 of the year.

Truite = 345

Truite, day 15 of Fructidor, day 75 of Summer, day 345 of the year.

Citron = 346

Citron, day 16 of Fructidor, day 76 of Summer, day 346 of the year.

Cardère = 347

Cardère, day 17 of Fructidor, day 77 of Summer, day 347 of the year.

Nerprun = 348

Nerprun, day 18 of Fructidor, day 78 of Summer, day 348 of the year.

Tagette = 349

Tagette, day 19 of Fructidor, day 79 of Summer, day 349 of the year.

Hotte = 350

Hotte, day 20 of Fructidor, day 80 of Summer, day 350 of the year.

Églantier = 351

Églantier, day 21 of Fructidor, day 81 of Summer, day 351 of the year.

Noisette = 352

Noisette, day 22 of Fructidor, day 82 of Summer, day 352 of the year.

Houblon = 353

Houblon, day 23 of Fructidor, day 83 of Summer, day 353 of the year.

Sorgho = 354

Sorgho, day 24 of Fructidor, day 84 of Summer, day 354 of the year.

Écrevisse = 355

Écrevisse, day 25 of Fructidor, day 85 of Summer, day 355 of the year.

Bigarade = 356

Bigarade, day 26 of Fructidor, day 86 of Summer, day 356 of the year.

VergeDOr = 357

Verge d’or, day 27 of Fructidor, day 87 of Summer, day 357 of the year.

Maïs = 358

Maïs, day 28 of Fructidor, day 88 of Summer, day 358 of the year.

Marron = 359

Marron, day 29 of Fructidor, day 89 of Summer, day 359 of the year.

Panier = 360

Panier, day 30 of Fructidor, day 90 of Summer, day 360 of the year.

LaFêteDeLaVertu = 361

La fête de la vertu, day 1 of Sans-culottides, day 361 of the year.

LaFêteDuGénie = 362

La fête du génie, day 2 of Sans-culottides, day 362 of the year.

LaFêteDuTravail = 363

La fête du travail, day 3 of Sans-culottides, day 363 of the year.

LaFêteDeLOpinion = 364

La fête de l’opinion, day 4 of Sans-culottides, day 364 of the year.

LaFêteDesRécompenses = 365

La fête des récompenses, day 5 of Sans-culottides, day 365 of the year.

LaFêteDeLaRévolution = 366

La fête de la Révolution, day 6 of Sans-culottides, day 366 of the year.

Implementations

impl DayOfYear

pub fn try_from_month_day(month: impl TryInto<Month>, day: Day) -> Option<Self>

Create a DayOfYear from a month (1-13) and day (1-30 for months 1-12, 1-6 for month 13).

pub fn into_month_day(self) -> (Month, Day)

Convert a DayOfYear into a (month, day) tuple.

pub fn is_sans_culottides(self) -> bool

Returns true if the day is in the Sans-culottides.

pub fn is_leap_day(self) -> bool

Returns true if the day is the leap day (Fête de la Révolution).

pub fn is_legal_for_year<L: LeapSystem>(self, year: Year) -> bool

Returns true if the day is legal for the given year in the given leap year system.