republican_calendar::date::leap

Trait LeapSystem

pub trait LeapSystem {
    const NAME: &'static str = "Unnamed Leap System";

    // Required method
    fn is_leap_year(year: Year) -> bool;

    // Provided methods
    fn leap_years_between(start: Year, end: Year) -> usize { ... }
    fn days_in_year(year: Year) -> usize { ... }
    fn days_in_years_between(start: Year, end: Year) -> isize { ... }
}

A trait of leap year systems for the Republican Calendar.

Provided Associated Constants

const NAME: &'static str = "Unnamed Leap System"

The name of the leap year system.

Required Methods

fn is_leap_year(year: Year) -> bool

Determine if a given year is a leap year.

Provided Methods

fn leap_years_between(start: Year, end: Year) -> usize

Count the number of leap years between two years (inclusive of start, exclusive of end).

fn days_in_year(year: Year) -> usize

Get the number of days in a given year.

fn days_in_years_between(start: Year, end: Year) -> isize

Count the number of days between two years (inclusive of start, exclusive of end).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.